+++ scripts/style_check_in_makefile.sh
@@ -46,16 +46,17 @@ function style_check
 }
 
 
+style_check src/lib/orionld/apiModel
 style_check src/lib/orionld/regCache
 style_check src/lib/orionld/forwarding
-style_check src/lib/orionld/prometheus
 style_check src/lib/orionld/mongoc
-style_check src/lib/orionld/payloadCheck
-style_check src/lib/orionld/serviceRoutines
 style_check src/lib/orionld/dbModel
-style_check src/lib/orionld/legacyDriver
 style_check src/lib/orionld/notifications
 style_check src/lib/orionld/common
+style_check src/lib/orionld/prometheus
+style_check src/lib/orionld/payloadCheck
+style_check src/lib/orionld/serviceRoutines
+style_check src/lib/orionld/legacyDriver
 style_check src/lib/orionld/q
 style_check src/lib/orionld/rest
 style_check src/lib/orionld/db
