+++ src/lib/orionld/kjTree/kjSysAttrsRemove.cpp
@@ -1,6 +1,3 @@
-#ifndef SRC_LIB_ORIONLD_KJTREE_KJTREELOG_H_
-#define SRC_LIB_ORIONLD_KJTREE_KJTREELOG_H_
-
 /*
 *
 * Copyright 2022 FIWARE Foundation e.V.
@@ -25,25 +22,31 @@
 *
 * Author: Ken Zangelin
 */
+#include <unistd.h>                                            // NULL
+
 extern "C"
 {
 #include "kjson/KjNode.h"                                      // KjNode
+#include "kjson/kjLookup.h"                                    // kjLookup
+#include "kjson/kjBuilder.h"                                   // kjChildRemove
 }
 
-
-
-// -----------------------------------------------------------------------------
-//
-// kjTreeLog -
-//
-#define kjTreeLog(tree, msg)             kjTreeLogFunction(tree, msg, __FILE__, __LINE__)
+#include "orionld/kjTree/kjSysAttrsRemove.h"                   // Own interface
 
 
 
 // -----------------------------------------------------------------------------
 //
-// kjTreeLogFunction -
+// kjSysAttrsRemove -
 //
-extern void kjTreeLogFunction(KjNode* tree, const char* msg, const char* fileName, int lineNo);
+void kjSysAttrsRemove(KjNode* container)
+{
+  KjNode* createdAtP  = kjLookup(container, "createdAt");
+  KjNode* modifiedAtP = kjLookup(container, "modifiedAt");
 
-#endif  // SRC_LIB_ORIONLD_KJTREE_KJTREELOG_H_
+  if (createdAtP != NULL)
+    kjChildRemove(container, createdAtP);
+
+  if (modifiedAtP != NULL)
+    kjChildRemove(container, modifiedAtP);
+}
