+++ src/lib/orionld/serviceRoutines/orionldGetEntity.cpp
@@ -43,462 +43,17 @@ extern "C"
 #include "orionld/mongoc/mongocEntityLookup.h"                   // mongocEntityLookup
 #include "orionld/dbModel/dbModelToApiEntity.h"                  // dbModelToApiEntity
 #include "orionld/kjTree/kjGeojsonEntityTransform.h"             // kjGeojsonEntityTransform
-#include "orionld/kjTree/kjChildCount.h"                         // kjChildCount
+#include "orionld/kjTree/kjSysAttrsRemove.h"                     // kjSysAttrsRemove
+#include "orionld/apiModel/ntonEntity.h"                         // ntonEntity
+#include "orionld/apiModel/ntosEntity.h"                         // ntosEntity
+#include "orionld/apiModel/ntocEntity.h"                         // ntocEntity
 #include "orionld/forwarding/ForwardPending.h"                   // ForwardPending
 #include "orionld/forwarding/regMatchForEntityGet.h"             // regMatchForEntityGet
 #include "orionld/forwarding/forwardingListsMerge.h"             // forwardingListsMerge
 #include "orionld/forwarding/forwardRequestSend.h"               // forwardRequestSend
 #include "orionld/forwarding/fwdPendingLookupByCurlHandle.h"     // fwdPendingLookupByCurlHandle
+#include "orionld/forwarding/fwdEntityMerge.h"                   // fwdEntityMerge
 #include "orionld/serviceRoutines/orionldGetEntity.h"            // Own interface
-
-
-
-// -----------------------------------------------------------------------------
-//
-// timestampMerge -
-//
-void timestampMerge(KjNode* apiEntityP, KjNode* additionP, KjNode* currentTsP, KjNode* newTsP, const char* tsName, bool newReplaces)
-{
-  if (currentTsP != NULL)  // This should always be the case
-  {
-    bool newIsNewer = strcmp(currentTsP->value.s, newTsP->value.s) < 0;
-
-    if (newReplaces == newIsNewer)
-      currentTsP->value.s = newTsP->value.s;
-  }
-  else
-  {
-    kjChildRemove(additionP, newTsP);
-    kjChildAdd(apiEntityP, newTsP);
-  }
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// newerAttribute -
-//
-KjNode* newerAttribute(KjNode* currentP, KjNode* pretenderP)
-{
-  KjNode* currentObservedAt   = kjLookup(currentP,   "observedAt");
-  KjNode* pretenderObservedAt = kjLookup(pretenderP, "observedAt");
-
-  if ((currentObservedAt == NULL) && (pretenderObservedAt == NULL))
-  {
-    KjNode* currentModifiedAt   = kjLookup(currentP,   "modifiedAt");
-    KjNode* pretenderModifiedAt = kjLookup(pretenderP, "modifiedAt");
-
-    if ((currentModifiedAt != NULL) && (pretenderModifiedAt != NULL))
-    {
-      if (strcmp(currentModifiedAt->value.s, pretenderModifiedAt->value.s) >= 0)
-        return currentP;
-      else
-        return pretenderP;
-    }
-    else if (currentModifiedAt != NULL)
-      return currentP;
-    else if (pretenderModifiedAt != NULL)
-      return pretenderP;
-    else
-      return currentP;  // Just pick one ...
-  }
-  else if (currentObservedAt == NULL)
-    return pretenderP;
-  else if (pretenderObservedAt == NULL)
-    return currentP;
-  else  // both non-NULL
-  {
-    if (strcmp(currentObservedAt->value.s, pretenderObservedAt->value.s) >= 0)
-      return currentP;
-    else
-      return pretenderP;
-  }
-
-  LM_W(("Not sure how we got here ... keeping the current - no replace"));
-  return currentP;
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// entityMerge -
-//
-bool entityMerge(KjNode* apiEntityP, KjNode* additionP, bool sysAttrs, bool auxiliary)
-{
-  KjNode* idP             = kjLookup(additionP,  "id");
-  KjNode* typeP           = kjLookup(additionP,  "type");
-
-  if (idP)
-    kjChildRemove(additionP, idP);
-  if (typeP)
-    kjChildRemove(additionP, typeP);
-
-  if (sysAttrs == true)  // sysAttrs for the final response
-  {
-    KjNode* addCreatedAtP   = kjLookup(additionP,  "createdAt");
-    KjNode* addModifiedAtP  = kjLookup(additionP,  "modifiedAt");
-    KjNode* createdAtP      = kjLookup(apiEntityP, "createdAt");
-    KjNode* modifiedAtP     = kjLookup(apiEntityP, "modifiedAt");
-
-    if (addCreatedAtP != NULL)
-      timestampMerge(apiEntityP, additionP, createdAtP,  addCreatedAtP,  "createdAt",  false);   // false: replace if older
-    if (addModifiedAtP != NULL)
-      timestampMerge(apiEntityP, additionP, modifiedAtP, addModifiedAtP, "modifiedAt", true);    // true: replace if newer
-  }
-
-  KjNode* attrP   = additionP->value.firstChildP;
-  KjNode* next;
-
-  while (attrP != NULL)
-  {
-    next = attrP->next;
-
-    KjNode* currentP = kjLookup(apiEntityP, attrP->name);
-
-    if (currentP == NULL)
-    {
-      kjChildRemove(additionP, attrP);
-      kjChildAdd(apiEntityP, attrP);
-    }
-    else if (auxiliary == false)  // two copies of the same attr ...  and NOT from an auxiliary registration
-    {
-      if (newerAttribute(currentP, attrP) == attrP)
-      {
-        kjChildRemove(apiEntityP, currentP);
-        kjChildRemove(additionP, attrP);
-        kjChildAdd(apiEntityP, attrP);
-      }
-    }
-
-    attrP = next;
-  }
-
-  return true;
-}
-
-
-
-#if 0
-static void attrsDebug(StringArray* attrV, const char* info)
-{
-  LM(("SLIST: %s", info));
-  if (attrV != NULL)
-  {
-    if (attrV->items == 0)
-    {
-      LM(("SLIST:   EMPTY attrV"));
-    }
-    else
-    {
-      for (int ix = 0; ix < attrV->items; ix++)
-      {
-        LM(("SLIST:   Attr %d: %s", ix, attrV->array[ix]));
-      }
-    }
-  }
-  else
-    LM(("SLIST:   NULL attrV"));
-}
-
-static void attrsDebug2(ForwardPending*  fwdPendingList)
-{
-  ForwardPending* fwdPendingP = fwdPendingList;
-
-  LM(("SLIST: *************** Entire ForwardPending list ***************************"));
-  while (fwdPendingP != NULL)
-  {
-    KjNode* regIdP = kjLookup(fwdPendingP->regP->regTree, "id");
-
-    LM(("SLIST: -------------------------------------------------------------"));
-    LM(("SLIST: Registration: %s", (regIdP != NULL)? regIdP->value.s : "unknown"));
-    LM(("SLIST: FwdOperation: %s", fwdOperations[fwdPendingP->operation]));
-    attrsDebug(fwdPendingP->attrList, "attrList");
-    LM(("SLIST: -------------------------------------------------------------"));
-
-    fwdPendingP = fwdPendingP->next;
-  }
-  LM(("SLIST: ******************************************************"));
-}
-#endif
-
-
-
-// -----------------------------------------------------------------------------
-//
-// sysAttrsRemove -
-//
-void sysAttrsRemove(KjNode* container)
-{
-  KjNode* createdAtP  = kjLookup(container, "createdAt");
-  KjNode* modifiedAtP = kjLookup(container, "modifiedAt");
-
-  if (createdAtP != NULL)
-    kjChildRemove(container, createdAtP);
-
-  if (modifiedAtP != NULL)
-    kjChildRemove(container, modifiedAtP);
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// langFixNormalized - from kjTreeFromContextAttribute.cpp
-//
-// FIXME: move to its own module
-//
-extern void langFixNormalized(KjNode* attrP, KjNode* typeP, KjNode* languageMapP, const char* lang);
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ntonSubAttribute -
-//
-void ntonSubAttribute(KjNode* saP, char* lang, bool sysAttrs)
-{
-  if (sysAttrs == false)
-    sysAttrsRemove(saP);
-
-  KjNode* typeP = kjLookup(saP, "type");
-  if ((lang != NULL) && (typeP != NULL) && (strcmp(typeP->value.s, "LanguageProperty")))
-  {
-    KjNode* languageMapP = kjLookup(saP, "languageMap");
-    langFixNormalized(saP, typeP, languageMapP, lang);
-  }
-
-  for (KjNode* fieldP = saP->value.firstChildP; fieldP != NULL; fieldP = fieldP->next)
-  {
-    if (strcmp(fieldP->name, "type")        == 0)  continue;
-    if (strcmp(fieldP->name, "value")       == 0)  continue;
-    if (strcmp(fieldP->name, "createdAt")   == 0)  continue;
-    if (strcmp(fieldP->name, "modifiedAt")  == 0)  continue;
-    if (strcmp(fieldP->name, "observedAt")  == 0)  continue;
-    if (strcmp(fieldP->name, "unitCode")    == 0)  continue;
-    if (strcmp(fieldP->name, "object")      == 0)  continue;
-    if (strcmp(fieldP->name, "languageMap") == 0)  continue;
-    if (strcmp(fieldP->name, "lang")        == 0)  continue;
-
-    ntonSubAttribute(fieldP, lang, sysAttrs);
-  }
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ntonAttribute -
-//
-void ntonAttribute(KjNode* attrP, char* lang, bool sysAttrs)
-{
-  //
-  // If ARRAY, we're dealing with datasetId ...  - later!
-  //
-  if (attrP->type == KjArray)
-  {
-    LM_E(("Multi-Attribute (datasetId) not supported, sorry ..."));
-    return;
-  }
-
-  if (sysAttrs == false)
-    sysAttrsRemove(attrP);
-
-  KjNode* typeP = kjLookup(attrP, "type");
-  if ((lang != NULL) && (typeP != NULL) && (strcmp(typeP->value.s, "LanguageProperty") == 0))
-  {
-    KjNode* languageMapP = kjLookup(attrP, "languageMap");
-    langFixNormalized(attrP, typeP, languageMapP, lang);
-  }
-
-  for (KjNode* fieldP = attrP->value.firstChildP; fieldP != NULL; fieldP = fieldP->next)
-  {
-    if (strcmp(fieldP->name, "type")        == 0)  continue;
-    if (strcmp(fieldP->name, "value")       == 0)  continue;
-    if (strcmp(fieldP->name, "createdAt")   == 0)  continue;
-    if (strcmp(fieldP->name, "modifiedAt")  == 0)  continue;
-    if (strcmp(fieldP->name, "observedAt")  == 0)  continue;
-    if (strcmp(fieldP->name, "unitCode")    == 0)  continue;
-    if (strcmp(fieldP->name, "object")      == 0)  continue;
-    if (strcmp(fieldP->name, "languageMap") == 0)  continue;
-    if (strcmp(fieldP->name, "lang")        == 0)  continue;
-
-    ntonSubAttribute(fieldP, lang, sysAttrs);
-  }
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ntonEntity - transform entity from normalized to fixed normalized
-//
-// Removing sysAttrs, fixing lang, ...
-//
-void ntonEntity(KjNode* apiEntityP, char* lang, bool sysAttrs)
-{
-  if (sysAttrs == false)
-    sysAttrsRemove(apiEntityP);
-
-  for (KjNode* fieldP = apiEntityP->value.firstChildP; fieldP != NULL; fieldP = fieldP->next)
-  {
-    if (strcmp(fieldP->name, "id")          == 0)  continue;
-    if (strcmp(fieldP->name, "type")        == 0)  continue;
-    if (strcmp(fieldP->name, "createdAt")   == 0)  continue;
-    if (strcmp(fieldP->name, "modifiedAt")  == 0)  continue;
-
-    // It's an attribute
-    ntonAttribute(fieldP, lang, sysAttrs);
-  }
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ntosEntity - normalized to simplified for an entity
-//
-// The combination Simplified+KeyValues isn't allowed, we'd never get this far
-//
-void ntosEntity(KjNode* apiEntityP, char* lang)
-{
-  KjNode* attrP = apiEntityP->value.firstChildP;
-  KjNode* next;
-
-  while (attrP != NULL)
-  {
-    next = attrP->next;
-
-    if (strcmp(attrP->name, "id") == 0)
-    {
-      attrP = next;
-      continue;
-    }
-    if (strcmp(attrP->name, "type") == 0)
-    {
-      attrP = next;
-      continue;
-    }
-
-    if (strcmp(attrP->name, "createdAt") == 0)
-    {
-      kjChildRemove(apiEntityP, attrP);
-      attrP = next;
-      continue;
-    }
-
-    if (strcmp(attrP->name, "modifiedAt") == 0)
-    {
-      kjChildRemove(apiEntityP, attrP);
-      attrP = next;
-      continue;
-    }
-
-    //
-    // It's an attribute
-    //
-    KjNode* valueP = kjLookup(attrP, "value");
-
-    if (valueP == NULL)
-      valueP = kjLookup(attrP, "object");
-
-    if (valueP == NULL)
-      valueP = kjLookup(attrP, "languageMap");
-
-    if (valueP != NULL)
-    {
-      attrP->value     = valueP->value;
-      attrP->type      = valueP->type;
-      attrP->lastChild = valueP->lastChild;
-    }
-    else
-      LM_E(("No attribute value (object/languageMap) found - this should never happen!!!"));
-
-    attrP = next;
-  }
-}
-
-
-
-// -----------------------------------------------------------------------------
-//
-// ntocEntity -
-//
-void ntocEntity(KjNode* apiEntityP, char* lang, bool sysAttrs)
-{
-  KjNode* attrP = apiEntityP->value.firstChildP;
-  KjNode* next;
-
-  while (attrP != NULL)
-  {
-    next = attrP->next;
-
-    if (strcmp(attrP->name, "id") == 0)
-    {
-      attrP = next;
-      continue;
-    }
-    if (strcmp(attrP->name, "type") == 0)
-    {
-      attrP = next;
-      continue;
-    }
-
-    if (strcmp(attrP->name, "createdAt") == 0)
-    {
-      if (sysAttrs == false)
-        kjChildRemove(apiEntityP, attrP);
-      attrP = next;
-      continue;
-    }
-
-    if (strcmp(attrP->name, "modifiedAt") == 0)
-    {
-      if (sysAttrs == false)
-        kjChildRemove(apiEntityP, attrP);
-      attrP = next;
-      continue;
-    }
-
-    //
-    // It's a regular attribute
-    //
-    // 1. Remove the attribute type
-    // 2. If only "value" left - simplified
-    //
-    KjNode* typeP       = kjLookup(attrP, "type");
-    KjNode* createdAtP  = kjLookup(attrP, "createdAt");
-    KjNode* modifiedAtP = kjLookup(attrP, "modifiedAt");
-
-    if (typeP != NULL)
-      kjChildRemove(attrP, typeP);
-
-    if (sysAttrs == false)
-    {
-      if (createdAtP != NULL)
-        kjChildRemove(attrP, createdAtP);
-
-      if (modifiedAtP != NULL)
-        kjChildRemove(attrP, modifiedAtP);
-    }
-
-    KjNode* valueP     = kjLookup(attrP, "value");
-    int     attrFields = kjChildCount(attrP);
-
-    if ((valueP != NULL) && (attrFields == 1))  // Simplified
-    {
-      attrP->value     = valueP->value;
-      attrP->type      = valueP->type;
-      attrP->lastChild = valueP->lastChild;
-    }
-    else  // Dig into sub-attrs - might want a recursive call here
-    {
-    }
-
-    attrP = next;
-  }
-}
 
 
 
@@ -663,7 +218,7 @@ bool orionldGetEntity(void)
           if (apiEntityP == NULL)
             apiEntityP = fwdPendingP->body;
           else
-            entityMerge(apiEntityP, fwdPendingP->body, orionldState.uriParamOptions.sysAttrs, fwdPendingP->regP->mode == RegModeAuxiliary);
+            fwdEntityMerge(apiEntityP, fwdPendingP->body, orionldState.uriParamOptions.sysAttrs, fwdPendingP->regP->mode == RegModeAuxiliary);
         }
         else
           LM_E(("Internal Error (parse error for the received response of a forwarded request)"));
